% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/command.R
\name{as.command}
\alias{as.command}
\title{Makes input as a commandline command.}
\usage{
as.command(cmd, character.only = FALSE, envir = as.environment(-1))
}
\arguments{
\item{cmd}{a command symbol, or a command string with \code{character.only = TRUE}}

\item{character.only}{assumes that \code{cmd} is given as a string}

\item{envir}{an environment that a command object to be assigned}
}
\description{
Makes input as a commandline command.
}

