% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loadCommand.R
\name{loadCommand}
\alias{loadCommand}
\title{Loads executable files in the specified directory}
\usage{
loadCommand(path, envir, inherits = FALSE)
}
\arguments{
\item{path}{a directory to search executable files}
}
\description{
Loads executable files in the specified directory
}

