% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/command.R
\name{print.command}
\alias{print.command}
\title{Invokes a command object as an OS command.}
\usage{
\method{print}{command}(command, intern = FALSE, ignore.stdout = FALSE,
  ignore.stderr = FALSE, wait = TRUE, input = NULL,
  show.output.on.console = TRUE, minimized = FALSE, invisible = TRUE)
}
\arguments{
\item{cmd}{a command object}
}
\value{
result of the command
}
\description{
Invokes a command object as an OS command.
}

