% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.R
\name{print.dev.null}
\alias{print.dev.null}
\title{Prints nothing.
This function is used for a comman that takes parameters}
\usage{
\method{print}{dev.null}(...)
}
\description{
Prints nothing.
This function is used for a comman that takes parameters
}

