% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/command.R
\name{[.command}
\alias{[.command}
\title{Use a command with commandline parameters.}
\usage{
\method{[}{command}(cmd, ..., input = NULL)
}
\arguments{
\item{cmd}{a command object}

\item{...}{commandline parameters}

\item{input}{input of command}
}
\value{
result of the command
}
\description{
Use a command with commandline parameters.
}

